requirements_old.txt
# Core framework - 2025 Latest Stable
fastapi>=0.112.0
uvicorn[standard]>=0.32.0
starlette>=0.40.0
python-dotenv>=1.1.1

# Async HTTP and utilities
httpx>=0.29.0
httpcore>=1.0.7
aiohttp>=3.11.0
tenacity>=9.0.0
anyio>=4.6.0
nest-asyncio>=1.6.0
python-multipart>=0.0.20
asgiref>=3.8.1
async-timeout>=5.0.1
aiohappyeyeballs>=2.6.0
aiosignal>=1.4.0

# Web scraping and sanitization
beautifulsoup4>=4.13.0
lxml>=5.0.0
bleach>=6.2.0

# PDF/image/OCR
PyMuPDF>=1.26.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.0

# AI, LLMs, and orchestration - 2025 Latest
openai>=1.54.0
anthropic>=0.37.0
transformers>=4.51.3
torch>=2.5.0
sentence-transformers>=3.2.0
huggingface-hub>=0.26.0
safetensors>=0.4.5
tqdm>=4.67.0
regex>=2024.11.6
langchain>=0.3.7
langchain-core>=0.3.17
langchain-text-splitters>=0.3.2
langgraph>=0.2.34
langgraph-checkpoint>=2.1.0
langgraph-prebuilt>=0.6.0
langgraph-sdk>=0.2.0
langsmith>=0.6.0
scikit-learn>=1.6.0
joblib>=1.4.2
ollama>=0.4.0  # 2025 API updates

# Retrieval: vector + keyword + graph - 2025 Latest
qdrant-client>=1.12.0
meilisearch>=1.5.0
python-arango>=8.2.0
chromadb>=0.5.20  # 2025 updates with improved performance

# Database
SQLAlchemy>=2.0.0
sqlalchemy-utils>=0.41.2
greenlet>=3.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Security & SSO
bcrypt>=4.3.0
cryptography>=42.0.0
PyJWT>=2.10.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
python-jose[cryptography]>=3.5.0

# Pydantic / Validation / Config - 2025 v2.x Latest
pydantic>=2.10.2
pydantic-settings>=2.8.2
pydantic_core>=2.30.1
typing_extensions>=4.12.2
annotated-types>=0.7.0

# Caching - 2025 Latest Redis
aiocache>=0.12.3
redis>=5.2.0
aioredis>=2.1.0

# Monitoring and Logging - 2025 Latest OpenTelemetry
prometheus_client>=0.21.0
structlog>=24.5.0
psutil>=6.1.0
opentelemetry-api>=1.28.2
opentelemetry-sdk>=1.28.2
opentelemetry-instrumentation>=0.49b2
opentelemetry-instrumentation-fastapi>=0.49b2
opentelemetry-instrumentation-logging>=0.49b2
opentelemetry-instrumentation-asgi>=0.49b2
opentelemetry-semantic-conventions>=0.49b2
opentelemetry-util-http>=0.49b2

# Dev & testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
websockets>=10.4

# Utils
click>=8.1.0
certifi>=2025.7.0
idna>=3.10
h11>=0.16.0
colorama>=0.4.6
sniffio>=1.1
multidict>=6.6.0
frozenlist>=1.7.0
attrs>=25.0.0
yarl>=1.20.0
packaging>=24.2.0
soupsieve>=2.7.0
typing-inspection>=0.4.0

# Network and Graph Analysis
networkx>=3.5.0

# Configuration and YAML
PyYAML>=6.0.0

# Data Analysis
pandas>=2.3.0
numpy>=2.3.0

# Advanced Analytics & Visualization
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=6.2.0
dash>=3.2.0
streamlit>=1.48.0
scipy>=1.16.0

# Multi-tenant & SSO Support
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# Additional utilities that might be needed
requests>=2.31.0
urllib3>=2.5.0
charset-normalizer>=3.4.0
